refactor: Split up player's handleInput function
The player's handleInput function was split up so that this.x and this.y get set within the update function since the comment's dictate that there should be an update function in the player class. This led to a bug where the user would continously get notified that the game was over.